Changes MNIST autocreate script to be executed from home directory, works on the conv_mnist example a little bit

+++ src/bin/create_mnist_subset.rs
@@ -1,26 +1,33 @@
 
 use std::fs;
+use std::path::Path;
 
 fn main() {
-    let MNIST_DIR = "../../data/mnist";
+    let MNIST_DIR = "data/mnist";
 
-    let TRAIN_SET_SIZE = 400;
+    let TRAIN_SET_SIZE = 1000;
     let TRAIN_DIR = MNIST_DIR.to_owned() + "/" + "train";
     let CONSOLIDATED_TRAIN_DIR = MNIST_DIR.to_owned() + "/" + "train_subset";
-    fs::remove_dir_all(&CONSOLIDATED_TRAIN_DIR).expect("Couldn't remove previous consolidated training directory");
+    if Path::new(&CONSOLIDATED_TRAIN_DIR).exists() {
+        fs::remove_dir_all(&CONSOLIDATED_TRAIN_DIR).expect("Couldn't remove previous consolidated training directory");
+    }
+    std::thread::sleep(std::time::Duration::from_millis(10));
     fs::create_dir_all(&CONSOLIDATED_TRAIN_DIR).expect("Couldn't build the consolidated training directory");
 
-    let TEST_SET_SIZE=100;
+    let TEST_SET_SIZE=200;
     let TEST_DIR = MNIST_DIR.to_owned() + "/" + "test";
     let CONSOLIDATED_TEST_DIR = MNIST_DIR.to_owned() + "/" + "test_subset";
-    fs::remove_dir_all(&CONSOLIDATED_TEST_DIR).expect("Couldn't remove previous consolidated training directory");
+    if Path::new(&CONSOLIDATED_TEST_DIR).exists() {
+        fs::remove_dir_all(&CONSOLIDATED_TEST_DIR).expect("Couldn't remove previous consolidated training directory");
+    }
     fs::create_dir_all(&CONSOLIDATED_TEST_DIR).expect("Couldn't build the consolidated testing directory");
 
+    // let group_names = vec!["zero","one","two","three","four","five","six","seven","eight","nine"];
+    let group_names = vec!["zero","one","two","three","four"];
 
+    // FOR THE TRAIN SUBSET
     // Let's accumulate the paths of all the images in the MNIST directory into their proper groups
     let paths = fs::read_dir(&TRAIN_DIR).unwrap();
-    // let group_names = vec!["zero","one","two","three","four","five","six","seven","eight","nine"];
-    let group_names = vec!["zero","one","two","three"];
     let mut groups: Vec<Vec<String>> = Vec::new();
     for group in &group_names {
         groups.push(vec![]);
@@ -44,9 +51,9 @@ fn main() {
         }
     }
 
+    // FOR THE TEST SUBSET
     // Let's accumulate the paths of all the images in the MNIST directory into their proper groups
     let paths = fs::read_dir(&TEST_DIR).unwrap();
-    let group_names = vec!["one","two","three","four","five","six","seven","eight","nine"];
     let mut groups: Vec<Vec<String>> = Vec::new();
     for group in &group_names {
         groups.push(vec![]);
